Consolidate python functionality into provider
