+++ src/providers/python.rs
@@ -1,9 +1,8 @@
-use anyhow::Result;
+use anyhow::{Result, Context};
 
 use crate::{
     nixpacks::{app::App, environment::Environment, nix::NixConfig},
-    python::pyproject,
-    Pkg,
+    Pkg, chain,
 };
 
 use super::Provider;
@@ -14,7 +13,7 @@ impl Provider for PythonProvider {
         "python"
     }
 
-    fn detect(&self, app: &crate::nixpacks::app::App, _env: &Environment) -> anyhow::Result<bool> {
+    fn detect(&self, app: &crate::nixpacks::app::App, _env: &Environment) -> Result<bool> {
         Ok(app.includes_file("main.py")
             || app.includes_file("requirements.txt")
             || app.includes_file("pyproject.toml"))
@@ -37,11 +36,11 @@ impl Provider for PythonProvider {
 
     fn suggested_start_command(&self, app: &App, _env: &Environment) -> Result<Option<String>> {
         if app.includes_file("pyproject.toml") {
-            if let Ok(meta) = pyproject::parse(app) {
+            if let Ok(meta) = parse_project(app) {
                 if let Some(entry_point) = meta.entry_point {
                     return match entry_point {
-                        pyproject::EntryPoint::Command(cmd) => Ok(Some(cmd)),
-                        pyproject::EntryPoint::Module(module) => {
+                        EntryPoint::Command(cmd) => Ok(Some(cmd)),
+                        EntryPoint::Module(module) => {
                             Ok(Some(format!("python -m {}", module)))
                         }
                     };
@@ -62,4 +61,71 @@ impl Provider for PythonProvider {
     ) -> Result<crate::nixpacks::nix::NixConfig> {
         Ok(NixConfig::new(vec![Pkg::new("pkgs.python38")]))
     }
+}
+
+#[allow(dead_code)]
+struct ProjectMeta {
+    pub project_name: Option<String>,
+    pub module_name: Option<String>,
+    pub entry_point: Option<EntryPoint>,
+}
+
+enum EntryPoint {
+    Command(String),
+    Module(String),
+}
+
+fn parse_project(app: &App) -> Result<ProjectMeta> {
+    if !app.includes_file("pyproject.toml") {
+        return Err(anyhow::anyhow!("no project.toml found"));
+    }
+    let pyproject: toml::Value = app
+        .read_toml("pyproject.toml")
+        .context("Reading pyproject.toml")?;
+    let project = pyproject.get("project");
+    let project_name = chain!(project =>
+        |proj| proj.get("name"),
+        |name| name.as_str(),
+        |name| Some(name.to_string())
+    );
+
+    let module_name = chain!(project =>
+        (
+            |proj| proj.get("packages"),
+            |pkgs| pkgs.as_array(),
+            |pkgs| pkgs.get(0),
+            |package| package.as_str(),
+            |name| Some(name.to_string())
+        );
+        (
+            |proj| proj.get("py-modules"),
+            |mods| mods.as_array(),
+            |mods| mods.get(0),
+            |module| module.as_str(),
+            |name| Some(name.to_string())
+        );
+        (
+            |_| project_name.to_owned()
+        )
+    );
+
+    let entry_point = chain!(project =>
+        (
+            |project| project.get("scripts"),
+            |scripts| scripts.as_table(),
+            |scripts| Some(scripts.keys()),
+            |mut cmds| cmds.next(),
+            |cmd| Some(EntryPoint::Command(cmd.to_string()))
+        );
+        (
+            |_| module_name.to_owned(),
+            |module| Some(EntryPoint::Module(module))
+        )
+    );
+
+    Ok(ProjectMeta {
+        project_name,
+        module_name,
+        entry_point,
+    })
 }
